todo-app/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
defusedxml==0.7.1
dj-rest-auth==4.0.0
Django==4.2.1
django-allauth==0.52.0
djangorestframework==3.14.0
idna==3.7
install==1.3.5
oauthlib==3.2.2
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
python3-openid==3.2.0
pytz==2024.1
requests==2.31.0
requests-oauthlib==2.0.0
sqlparse==0.5.0
urllib3==2.2.1
